refactor: component 구조 코드리뷰
