+++ src/components/bookcase/index.ts
@@ -1,4 +1,4 @@
-export { default as BookcaseContainer } from "./BookcaseContainer";
+export { default as BookcaseWrapper } from "./BookcaseWrapper";
 export { default as Navigation } from "./Navigation";
 export { default as PeriRead } from "./PeriRead";
 export { default as PostRead } from "./PostRead";
