+++ src/components/common/MainHeader.tsx
@@ -1,8 +1,8 @@
-interface Children {
+interface MainHeaderProps {
   children: string;
 }
 
-export default function Header(props: Children) {
+export default function MainHeader(props: MainHeaderProps) {
   const children = props.children;
 
   return <header>{children}</header>;
