+++ src/components/common/index.ts
@@ -1,4 +1,4 @@
 export { default as Button } from "./Button";
-export { default as Header } from "./Header";
-export { default as LogoContainer } from "./LogoContainer";
-export { default as NavContainer } from "./NavContainer";
+export { default as MainHeader } from "./MainHeader";
+export { default as NavHeader } from "./NavHeader";
+export { default as NavWrapper } from "./NavWrapper";
