+++ src/components/main/index.ts
@@ -1,2 +1,2 @@
 export { default as Banner } from "./Banner";
-export { default as MainContainer } from "./MainContainer";
+export { default as MainWrapper } from "./MainWrapper";
