+++ src/pages/Bookcase.tsx
@@ -1,11 +1,11 @@
-import { BookcaseContainer } from "../components/bookcase";
-import { NavContainer } from "../components/common";
+import { BookcaseWrapper } from "../components/bookcase";
+import { NavWrapper } from "../components/common";
 
 export default function Bookcase() {
   return (
     <>
-      <NavContainer />
-      <BookcaseContainer />
+      <NavWrapper />
+      <BookcaseWrapper />
     </>
   );
 }
