+++ src/pages/index.ts
@@ -1,4 +1,4 @@
 export { default as Bookcase } from "./Bookcase";
 export { default as Login } from "./Login";
 export { default as Main } from "./Main";
-export { default as SignUp } from "./SignUp";
+export { default as Signup } from "./Signup";
